File chunks saved and counted
I got pretty far on the resumable upload today. Cribbing from Barbara's WARC uploader in the partner-site, my branch handles receiving chunked-files, saving the chunks to a org-specific temp dir, checking all of the chunks have been received for a file, as well as two new models for the DB: Deposit and DepositFile.

I'm thinking when a user initiates a deposit, we create a record in the Deposit table. That record has a state machine to keep track of REGISTERED -> UPLOADED -> HASHED -> REPLICATED. Either as each chunk gets sent, or in one payload we also create DepositFile records, storing the flow.js metadata for the file, and the same state machine as for the Deposit. I'm not sure yet if the DepositFile table needs to be distinct from the main file table you're making. But I'm using it for now to get it working.

When the chunk post handler sees that all of the required chunks for a file have been received, it moves the state of the DepositFile model from REGISTERED -> UPLOADED. It then returns. This will allow us to take the file hashing out of the request/response cycle, which could be causing timeouts. A separate process, running as a Django Admin command on a cron could look DepositFiles in UPLOADED state, acquire a lock on that record, get the hash of the file over all of its chunks, check to see if that file exists in the local copy of hashFS, then merge the chunks into that file with the hash as the filename. Once thats done, update the DepositFile state from UPLOADED -> HASHED.

Perhaps a separate process, or maybe the same process gets DepositFiles in HASHED state, looks up what are all of the replication targets for this file (another server, HDFS, petabox), and works on moving them.
Somewhere in here we also build the FileNode record.

Once all of DepositFiles of a given deposit have reached a state we update the state of the Deposit. Or another way, the state of the Deposit is the minimum state of its files.

flow.js gives good hooks for tracking the transfer progress of all of the components of an upload. The UI will need another method that, I suppose, polls the app for updates to the hashing and replication processes.

I don't want users deleting data from their systems the moment they see the file has finished transferring. Until the file is REPLICATED, it's at (small but real) risk on our side.