fsync chunks, prevent race condition, update deposit_file
Sam suggested we make liberal use of fsync for file uploads. Now we save chunks to an even more temporary path, flush, fsync, then move them to the path that the "do we have all the chunks for this file" logic expects. This fixes an issue with parallel uploads where a chunk will be partially written, and the code will assume all chunks are present then throw an error when it runs the file size comparison sanity check.

After we check all chunks are present, we update the DepositFile record to state UPLOADED. I've tested the process_chunked_files utility picks it up from there.