handle flow.js test for chunk
flow.js sends a GET request to see if we already have a given chunk. Reply with status=200, and it will skip POSTing the chunk. Reply with status=204, and it will POST the chunk.

The approach taken in this commit can lead to skipping all chunks for a DepositFile, and that DepositFile staying in state=REGISTERED.