TreeNode parent_id+name and sha256_sum indexes
Accessing TreeNode by a parent_id and the name of one its children is a
common pattern. We'll start at some root node (Org or Collection) then
iteratively follow a given path/to/some/node.

Here an index and unique constraint is added on parent_id+name. This
makes the "human" (not ltree) path/to/some/node unambiguous and should
be a faster lookup.

An index on TreeNode.sha256_sum is also added in anticipation of lookups
for garbage collection and deduplication.