api.collections_stats uses TreeNode
The whole dashboard API needs to be re-thought, but for now raw SQL can
be tortured into producing a report with the expected shape.

Annotating aggregates with the Django ORM almost works, but it's
designed to work with foreign key relationships, not ltree descendant
relationships. To get the total file size of each collection in one
query I had to switch to raw SQL.

It's a little tricky dealing with Collection and TreeNode being
different tables. Knowing a Collection.id isn't enough, we need the
Collection.tree_node.path.